+++ ext/tensor/decompositions/cholesky.zep.c
@@ -192,7 +192,7 @@ PHP_METHOD(Tensor_Decompositions_Cholesky, decompose) {
 	object_init_ex(return_value, tensor_decompositions_cholesky_ce);
 	ZEPHIR_CALL_CE_STATIC(&_29, tensor_matrix_ce, "quick", &_30, 0, &l);
 	zephir_check_call_status();
-	ZEPHIR_CALL_METHOD(NULL, return_value, "__construct", NULL, 24, &_29);
+	ZEPHIR_CALL_METHOD(NULL, return_value, "__construct", NULL, 23, &_29);
 	zephir_check_call_status();
 	RETURN_MM();