+++ optimizers/ArgmaxOptimizer.php
@@ -1,0 +1,72 @@
+<?php
+
+namespace Zephir\Optimizers\FunctionCall;
+
+use Zephir\Call;
+use Zephir\CompilationContext;
+use Zephir\CompiledExpression;
+use Zephir\CompilerException;
+use Zephir\HeadersManager;
+use Zephir\Optimizers\OptimizerAbstract;
+
+class ArgmaxOptimizer extends OptimizerAbstract
+{
+    /**
+     * @param array $expression
+     * @param Call $call
+     * @param CompilationContext $context
+     * @throws CompilerException
+     * @return bool|CompiledExpression
+     */
+    public function optimize(array $expression, Call $call, CompilationContext $context)
+    {
+        if (!isset($expression['parameters'])) {
+            return false;
+        }
+
+        if (count($expression['parameters']) !== 1) {
+            throw new CompilerException(
+                'Argmax accepts one and only one argument.',
+                $expression
+            );
+        }
+
+        $call->processExpectedReturn($context);
+
+        $symbolVariable = $call->getSymbolVariable();
+
+        if ($symbolVariable->getType() !== 'variable') {
+            throw new CompilerException(
+                'Returned values by functions can only be assigned to variant variables.',
+                $expression
+            );
+        }
+
+        if ($call->mustInitSymbolVariable()) {
+            $symbolVariable->initVariant($context);
+        }
+
+        $context->headersManager->add(
+            'include/indexing',
+            HeadersManager::POSITION_LAST
+        );
+
+        $resolvedParams = $call->getResolvedParams(
+            $expression['parameters'],
+            $context,
+            $expression
+        );
+
+        $symbol = $context->backend->getVariableCode($symbolVariable);
+
+        $context->codePrinter->output(
+            'tensor_argmax(' . $symbol . ', ' . $resolvedParams[0] . ');'
+        );
+
+        return new CompiledExpression(
+            'variable',
+            $symbolVariable->getRealName(),
+            $expression
+        );
+    }
+}