+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Individual Arithmetic and Comparison methods now public
     - Matrix reductions have their own namespace
     - Optimized Matrix structural operations
+    - Added argmin and argmax custom optimizers
 
 - 2.0.3
     - Matrix decomp methods now return decomp objects